Move dashboard content to dashboard page
Add a home page with placeholder content.
Update navbar links
